requirements.txt
numpy==1.23.4
pathtools==0.1.2
Pillow==9.3.0
PyYAML==6.0
torch==1.13.0
torchvision==0.14.0
urllib3==1.26.12
pandas
scikit-iamge
